requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
google-genai
py-autoclean